chore: create SheetModals.jsx
